+++ src/apps/seelenweg/modules/item/app/PinnedApp.ts
@@ -1,21 +1,25 @@
 import { convertFileSrc, invoke } from '@tauri-apps/api/core';
-import { PinnedAppWegItem, SwItemType } from 'seelen-core';
+import { PinnedWegItem, SwItemType } from 'seelen-core';
 
 import { LAZY_CONSTANTS } from '../../shared/utils/infra';
 
-import { ExtendedPinnedAppWegItem } from '../../shared/store/domain';
+import { ExtendedPinnedWegItem } from '../../shared/store/domain';
+
+export function isWindowsLink(item: PinnedWegItem): boolean {
+  return item.path.toLowerCase().endsWith('.lnk');
+}
 
 export class SwPinnedAppUtils {
-  static async fromSaved(item: PinnedAppWegItem): Promise<ExtendedPinnedAppWegItem> {
+  static async fromSaved(item: PinnedWegItem): Promise<ExtendedPinnedWegItem> {
     let icon_path =
-      (await invoke<string | null>('get_icon', { path: item.execution_path })) ||
-      LAZY_CONSTANTS.MISSING_ICON_PATH;
+      (await invoke<string | null>('get_icon', {
+        path: isWindowsLink(item) ? item.path : item.execution_command,
+      })) || LAZY_CONSTANTS.MISSING_ICON_PATH;
 
     return {
-      type: SwItemType.PinnedApp,
+      ...item,
+      type: SwItemType.Pinned,
       icon: convertFileSrc(icon_path),
-      exe: item.exe,
-      execution_path: item.execution_path,
       title: '',
       opens: [],
     };