+++ src/background/exposed.rs
@@ -40,12 +40,15 @@ fn select_file_on_explorer(path: String) -> Result<()> {
 }
 
 #[tauri::command(async)]
-fn open_file(path: String) -> Result<()> {
-    get_app_handle()
+fn open_file(path: String, args: Option<String>) -> Result<()> {
+    let mut command = get_app_handle()
         .shell()
         .command("cmd")
-        .args(["/c", "explorer", &path])
-        .spawn()?;
+        .args(["/C", "start", "", &path]);
+    if let Some(arg) = args {
+        command = command.arg(&arg);
+    }
+    command.spawn()?;
     Ok(())
 }